added laps, change data sync to 2 months

+++ database/migrations/2018_04_16_075057_create_activities_table.php
@@ -15,7 +15,7 @@ class CreateActivitiesTable extends Migration
   {
     Schema::create('activities', function (Blueprint $table) {
       $table->increments('id');
-      $table->integer('strava_id')->unsigned()->unique();
+      $table->bigInteger('strava_id')->unsigned()->unique();
       $table->integer('user_id')->unsigned();
       $table->string('name');
       $table->double('distance')->unsigned();
